Modified User, UserDTO and setup home page
UserManager is now using UserDTO as return type instead of User.

Modified User; removed all the transient values so it will reflect only
the fields in the database.

Transferred UserDetails interface to the UserDTO as it will be the
model to be used by the implementing class of UserDetailService.

Has briefly started the redirection of home page based on the role